moved dashboard-zero onto it's own folder
