+++ data/repos.csv
@@ -115,7 +115,7 @@ django-session-csrf,74,26,11,Python
 django-jobvite,11,7,0,Python
 domesday,5,3,0,Python
 server-core,3,2,0,Python
-pdf.js,13103,3061,582,JavaScript
+pdf.js,13104,3061,582,JavaScript
 sumo-tests,34,50,7,Python
 spark-eol,0,2,1,JavaScript
 server-share,6,2,0,Python
@@ -830,7 +830,7 @@ forums.mozilla.org,2,1,0,Shell
 wiki.mozilla.org,8,10,1,PHP
 firefox10.org,4,5,1,null
 node-firefox-find-ports,11,6,4,JavaScript
-firefox-ios,5018,891,94,C
+firefox-ios,5019,891,94,C
 fxa-relier-client,2,5,4,JavaScript
 fxa-local-dev,21,10,0,JavaScript
 marketplace-core-modules,6,11,0,JavaScript
